Redesigned drain hole plug with larger 5mm wedge holes.
These should hopefully not clog up that easily with the hair and toilet
paper that sometimes get washed into the drainage hole. Problem is,
smaller cockroaches and insects can make it through the hole pretty
easily. We shall see if that would be a problem.

Also added a fillet to the base, as new prints are difficult to insert
into the hole.