Added PVC pipe endcap.
I wanted a flush endcap for my 20mm PVC electrical fittings, either a
pipe or a junction box. Pipes are 20mm OD, so the junction boxes must
accept these.  They can either be fully sealed, or have a small gap for
a pair of 1.5mm wires to come through.

They also have a small lip on either side for a flat-head screwdriver to
be used for prying them out.