Added torch tailcap pusher for replacing tail switches.
This spacer allows you to use a clamp/vice to push the tailcap switch
into the tailcap itself without trying to hammer it in. Switch OD is
approx 20mm.

It didn't work for me though, I think I might have needed to sand the
switch assembly closer to the inner diameter of the aluminium cap before
trying to wedge it in.