Added Panasonic MC-CG381 vacuum hose ring
It's a plastic ring that holds the vacuum hose to the accessory
handpiece. For some reason the plastic clip on this ring broke off quite
early on in its life and the hose was unable to be securely attached to
the handpiece. Just recently had to use this vacuum again, so I designed
a replacement hose ring.